gametalk_api/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.2.2
django-filter==23.2
Pillow==10.0.0